Extract FeedLoaderStub to its own file
