+++ EssentialApp/EssentialAppTests/FeedLoaderStub.swift
@@ -1,0 +1,19 @@
+//
+//  FeedLoaderStub.swift
+//  EssentialApp
+//
+//  Created by Anthony on 20/7/25.
+//
+import EssentialFeed
+
+class FeedLoaderStub: FeedLoader {
+    private let result: FeedLoader.Result
+    
+    init(result: FeedLoader.Result) {
+        self.result = result
+    }
+    
+    func load(completion: @escaping (FeedLoader.Result) -> Void) {
+        completion(result)
+    }
+}
